requirements.cpu.txt
-r ./requirements.base.txt
-f https://data.pyg.org/whl/torch-1.12.0+cpu.html
torch-scatter==2.1.0
torch-sparse==0.6.15
torch-cluster==1.6.0
torch-spline-conv==1.2.1
torch-geometric~=2.2.0